Refactor Lookup Excel functions - LOOKUP(), VLOOKUP() and HLOOKUP() - into a separate class
